+++ examples/readme_example.py
@@ -1,5 +1,5 @@
-from population_modeling import *
-from population_modeling.statistic import Statistics
+from src.population_modeling import *
+from src.stats.statistic import Statistics
 
 first_bacteria = create_bacteria(p_for_death=0.1, p_for_reproduction=0.8)  # creating first bacteria to start a population
 population = create_population(first_bacteria)  # creating population
