--- population_modeling/__init__.py
@@ -1,6 +1,0 @@
-from population_modeling.population import create_population
-from population_modeling.selector.selector import *
-from population_modeling.selector.selector_params import *
-from population_modeling.bacteria import *
-from population_modeling.mutations.normal_mutator import *
-from population_modeling.life_cycle import LifeCycle
