+++ src/population_modeling/__init__.py
@@ -1,0 +1,5 @@
+from .mutations.normal_mutator import *
+from .life_cycle import LifeCycle
+from .bacteria import *
+from .population import *
+from .selector.selector import *
