+++ src/ui/callbacks/abstract_callback.py
@@ -1,0 +1,14 @@
+from abc import abstractmethod
+
+
+class AbstractCallback:
+    def __init__(self, params):
+        self.params = params
+
+    def get_params(self):    # [INPUT, OUTPUT, ...]
+        raise NotImplementedError
+
+    @abstractmethod
+    def get_function(self):
+        # CAllback implementation
+        raise NotImplementedError
