+++ src/ui/pages/abstract_page.py
@@ -1,0 +1,11 @@
+from abc import abstractmethod
+
+
+class AbstractPage:
+    @abstractmethod
+    def get_layout(self):
+        raise NotImplementedError
+
+    @abstractmethod
+    def get_callbacks(self):
+        raise NotImplementedError
